backend/requirements.txt
starlette==0.37.2
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
